+++ lib/src/providers/auth.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../flutter_login.dart';
+import 'package:flutter_login/flutter_login.dart';
 
 enum AuthMode { signup, login }
 
@@ -17,7 +17,12 @@ typedef SignupCallback = Future<String?>? Function(SignupData);
 /// The additional fields are provided as an `HashMap<String, String>`
 /// The result is an error message, callback successes if message is null
 typedef AdditionalFieldsCallback = Future<String?>? Function(
-    Map<String, String>);
+  Map<String, String>,
+);
+
+/// A callback which can be used to check data before switching
+/// The result is an error message, callback successes if message is null
+typedef BeforeAdditionalFieldsCallback = Future<String?>? Function(SignupData);
 
 /// If the callback returns true, the additional data card is shown
 typedef ProviderNeedsSignUpCallback = Future<bool> Function();
@@ -38,21 +43,21 @@ typedef ConfirmSignupCallback = Future<String?>? Function(String, LoginData);
 typedef ConfirmRecoverCallback = Future<String?>? Function(String, LoginData);
 
 class Auth with ChangeNotifier {
-  Auth(
-      {this.loginProviders = const [],
-      this.onLogin,
-      this.onSignup,
-      this.onRecoverPassword,
-      this.onConfirmRecover,
-      this.onConfirmSignup,
-      this.onResendCode,
-      String email = '',
-      String password = '',
-      String confirmPassword = '',
-      String verificationCode = '',
-      AuthMode initialAuthMode = AuthMode.login,
-      this.termsOfService = const []})
-      : _email = email,
+  Auth({
+    this.loginProviders = const [],
+    this.onLogin,
+    this.onSignup,
+    this.onRecoverPassword,
+    this.onConfirmRecover,
+    this.onConfirmSignup,
+    this.onResendCode,
+    this.beforeAdditionalFieldsCallback,
+    String email = '',
+    String password = '',
+    String confirmPassword = '',
+    AuthMode initialAuthMode = AuthMode.login,
+    this.termsOfService = const [],
+  })  : _email = email,
         _password = password,
         _confirmPassword = confirmPassword,
         _verificationCode = verificationCode,
@@ -66,6 +71,7 @@ class Auth with ChangeNotifier {
   final ConfirmSignupCallback? onConfirmSignup;
   final SignupCallback? onResendCode;
   final List<TermOfService> termsOfService;
+  final BeforeAdditionalFieldsCallback? beforeAdditionalFieldsCallback;
 
   AuthType _authType = AuthType.userPassword;
 
@@ -137,7 +143,7 @@ class Auth with ChangeNotifier {
 
   List<TermOfServiceResult> getTermsOfServiceResults() {
     return termsOfService
-        .map((e) => TermOfServiceResult(term: e, accepted: e.getStatus()))
+        .map((e) => TermOfServiceResult(term: e, accepted: e.checked))
         .toList();
   }
 }