+++ lib/src/widgets/term_of_service_checkbox.dart
@@ -7,10 +7,10 @@ class TermCheckbox extends StatefulWidget {
   final bool validation;
 
   const TermCheckbox({
-    Key? key,
+    super.key,
     required this.termOfService,
     this.validation = true,
-  }) : super(key: key);
+  });
 
   @override
   State<TermCheckbox> createState() => _TermCheckboxState();
@@ -20,15 +20,16 @@ class _TermCheckboxState extends State<TermCheckbox> {
   @override
   Widget build(BuildContext context) {
     return CheckboxFormField(
-      onChanged: (value) => widget.termOfService.setStatus(value!),
+      onChanged: (value) {
+        widget.termOfService.checked = value ?? false;
+      },
       initialValue: widget.termOfService.initialValue,
       title: widget.termOfService.linkUrl != null
           ? InkWell(
               onTap: () {
-                launchUrl(Uri(path: widget.termOfService.linkUrl!));
+                launchUrl(Uri.parse(widget.termOfService.linkUrl!));
               },
               child: Row(
-                crossAxisAlignment: CrossAxisAlignment.center,
                 children: [
                   Flexible(
                     child: Text(
@@ -56,7 +57,7 @@ class _TermCheckboxState extends State<TermCheckbox> {
       validator: (bool? value) {
         if (widget.validation &&
             widget.termOfService.mandatory &&
-            !widget.termOfService.getStatus()) {
+            !widget.termOfService.checked) {
           return widget.termOfService.validationErrorMessage;
         }
         return null;
@@ -66,36 +67,32 @@ class _TermCheckboxState extends State<TermCheckbox> {
 }
 
 class CheckboxFormField extends FormField<bool> {
-  CheckboxFormField(
-      {Key? key,
-      required Widget title,
-      required FormFieldValidator<bool> validator,
-      String validationErrorMessage = '',
-      bool initialValue = false,
-      bool autoValidate = true,
-      required ValueChanged<bool?> onChanged})
-      : super(
-            key: key,
-            validator: validator,
-            initialValue: initialValue,
-            builder: (FormFieldState<bool> state) {
-              return CheckboxListTile(
-                dense: true,
-                title: title,
-                value: state.value,
-                onChanged: (value) {
-                  onChanged(value);
-                  state.didChange(value);
-                },
-                subtitle: state.hasError
-                    ? Builder(
-                        builder: (BuildContext context) => Text(
-                          state.errorText!,
-                          style: TextStyle(color: Theme.of(context).errorColor),
-                        ),
-                      )
-                    : null,
-                controlAffinity: ListTileControlAffinity.leading,
-              );
-            });
+  CheckboxFormField({
+    super.key,
+    required Widget title,
+    required FormFieldValidator<bool> super.validator,
+    bool super.initialValue = false,
+    required ValueChanged<bool?> onChanged,
+  }) : super(
+          builder: (FormFieldState<bool> state) {
+            return CheckboxListTile(
+              dense: true,
+              title: title,
+              value: state.value,
+              onChanged: (value) {
+                onChanged(value);
+                state.didChange(value);
+              },
+              subtitle: state.hasError
+                  ? Builder(
+                      builder: (BuildContext context) => Text(
+                        state.errorText!,
+                        style: TextStyle(color: Theme.of(context).errorColor),
+                      ),
+                    )
+                  : null,
+              controlAffinity: ListTileControlAffinity.leading,
+            );
+          },
+        );
 }